refactor: seperate SessionHandler from commandHandler
