+++ server/src/api/socket.js
@@ -18,10 +18,10 @@ function removeFromSession(sessionId, event, callback) {
   )
 }
 
-export function pushToSession(sessionId, ...args) {
+export function pushToSession(sessionId, event, ...args) {
   const callbacks = sessions.get(sessionId)
   if (callbacks) {
-    callbacks.forEach((callback) => callback(...args))
+    callbacks.forEach((callback) => callback(event, ...args))
   }
 }
 
@@ -68,8 +68,8 @@ async function connectClient(socket) {
     stream.socket.write(message)
   })
 
-  const socketCallback = (event, data) => {
-    socket.emit(event, data)
+  const socketCallback = (event, ...data) => {
+    socket.emit(event, ...data)
   }
 
   listenToSession(session.id, socketCallback)
