+++ server/src/game/commandHandler.js
@@ -2,6 +2,7 @@ import minimist from 'minimist'
 import { FileGraph } from '@linux-odyssey/file-graph'
 import Quest from '../models/quest.js'
 import { pushToSession } from '../api/socket.js'
+import SessionHandler from './sessionHandler.js'
 
 // 檢查 pattern 是否符合 input
 const checkMatch = (pattern, input) => {
@@ -12,9 +13,9 @@ const checkMatch = (pattern, input) => {
   })
 }
 
-export default class CommandHandler {
+export default class CommandHandler extends SessionHandler {
   constructor(session, commandInput, additionalData) {
-    this.session = session
+    super(session)
     this.commandInput = commandInput
     this.argv = minimist(this.commandInput.command.split(' '))
 
@@ -52,54 +53,11 @@ export default class CommandHandler {
     this.session.graph = graph
   }
 
-  getTask(stage) {
-    return this.session.tasks.find((t) => t.id === stage.id)
-  }
-
-  isCompleted(stage) {
-    const task = this.getTask(stage)
-    return task ? task.completed : false
-  }
-
-  getStages() {
-    return this.quest.stages.filter(
-      (stage) =>
-        stage.requirements.every((r) => this.isCompleted({ id: r })) &&
-        (stage.repeatable || !this.isCompleted(stage))
-    )
-  }
-
   isMatch(stage) {
     const keys = ['command', 'output', 'error', 'pwd']
     return keys.every((k) =>
       checkMatch(stage.condition[k], this.commandInput[k])
     )
-  }
-
-  execute(stage) {
-    this.session.tasks
-      .filter((t) => t.id === stage.id)
-      .forEach((t) => {
-        t.completed = true
-      })
-    this.session.hints.push(...stage.hints)
-    if (stage.id === 'END') {
-      this.session.status = 'finished'
-      this.session.finishedAt = new Date()
-    }
-    const stages = this.getStages()
-    console.debug(stages)
-    const newTasks = stages
-      .filter((s) => s.task)
-      .map((s) => ({ id: s.id, name: s.task, completed: false }))
-
-    this.session.tasks.push(...newTasks)
-    pushToSession(this.session.id, 'tasks', this.session.tasks)
-    return {
-      responses: stage.responses,
-      hints: stage.hints,
-      tasks: newTasks,
-    }
   }
 
   async run() {
